+++ packages/gatsby-source-medusa/src/utils/capitalize.ts
@@ -1,0 +1,3 @@
+export function capitalize(s: string): string {
+  return s[0].toUpperCase() + s.slice(1)
+}