+++ packages/medusa/src/api/routes/admin/discounts/remove-valid-product.ts
@@ -1,7 +1,7 @@
-import DiscountService from "../../../../services/discount"
 import { defaultAdminDiscountsFields, defaultAdminDiscountsRelations } from "."
+import DiscountService from "../../../../services/discount"
 /**
- * @oas [post] /discounts/{id}/products/{product_id}
+ * @oas [delete] /discounts/{id}/products/{product_id}
  * operationId: "DeleteDiscountsDiscountProductsProduct"
  * summary: "Remove Product availability"
  * description: "Removes a Product from the list of Products that a Discount can be used for."
@@ -22,10 +22,10 @@ import { defaultAdminDiscountsFields, defaultAdminDiscountsRelations } from "."
  *               $ref: "#/components/schemas/discount"
  */
 export default async (req, res) => {
-  const { discount_id, variant_id } = req.params
+  const { discount_id, product_id } = req.params
 
   const discountService: DiscountService = req.scope.resolve("discountService")
-  await discountService.removeValidProduct(discount_id, variant_id)
+  await discountService.removeValidProduct(discount_id, product_id)
 
   const discount = await discountService.retrieve(discount_id, {
     select: defaultAdminDiscountsFields,