+++ packages/medusa/src/api/routes/admin/regions/set-metadata.ts
@@ -4,8 +4,8 @@ import RegionService from "../../../../services/region"
 import { defaultAdminRegionFields, defaultAdminRegionRelations } from "."
 
 /**
- * @oas [delete] /regions/{id}/payment-providers/{provider_id}
- * operationId: "PostRegionsRegionPaymentProvidersProvider"
+ * @oas [post] /regions/{id}/metadata
+ * operationId: "PostRegionsRegionMetadata"
  * summary: "Set the metadata of a Region"
  * description: "Sets the metadata of a Region"
  * x-authenticated: true
@@ -28,10 +28,7 @@ import { defaultAdminRegionFields, defaultAdminRegionRelations } from "."
 export default async (req, res) => {
   const { id } = req.params
 
-  const validated = await validator(
-    AdminPostRegionsRegionPaymentProvidersProviderReq,
-    req.body
-  )
+  const validated = await validator(AdminPostRegionsRegionMetadata, req.body)
 
   const regionService: RegionService = req.scope.resolve("regionService")
   await regionService.setMetadata(id, validated.key, validated.value)
@@ -44,7 +41,7 @@ export default async (req, res) => {
   res.status(200).json({ region })
 }
 
-export class AdminPostRegionsRegionPaymentProvidersProviderReq {
+export class AdminPostRegionsRegionMetadata {
   @IsString()
   key: string